Extract util packages to separate module
