+++ src/main/java/io/luna/LunaServer.java
@@ -7,9 +7,9 @@ import io.luna.game.plugin.PluginBootstrap;
 import io.luna.net.LunaChannelFilter;
 import io.luna.net.LunaChannelInitializer;
 import io.luna.net.msg.GameMessageRepository;
-import io.luna.util.AsyncExecutor;
-import io.luna.util.ThreadUtils;
-import io.luna.util.Tuple;
+import io.luna.util.concurrent.AsyncExecutor;
+import io.luna.util.concurrent.ThreadUtils;
+import io.luna.util.common.Tuple;
 import io.luna.util.parser.impl.BlacklistFileParser;
 import io.luna.util.parser.impl.EquipmentDefinitionFileParser;
 import io.luna.util.parser.impl.ItemDefinitionFileParser;
