+++ src/main/java/io/luna/game/plugin/PluginBootstrap.java
@@ -7,9 +7,9 @@ import io.luna.game.event.EventListener;
 import io.luna.game.event.EventListenerPipelineSet;
 import io.luna.game.event.EventMatcherListener;
 import io.luna.game.service.GameService;
-import io.luna.util.AsyncExecutor;
-import io.luna.util.ThreadUtils;
-import io.luna.util.Tuple;
+import io.luna.util.concurrent.AsyncExecutor;
+import io.luna.util.concurrent.ThreadUtils;
+import io.luna.util.common.Tuple;
 import io.luna.util.gui.PluginGui;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
@@ -19,7 +19,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
