+++ src/main/java/io/luna/game/service/LogoutService.java
@@ -10,7 +10,7 @@ import org.apache.logging.log4j.message.ParameterizedMessage;
 
 import java.util.Set;
 
-import static io.luna.util.ThreadUtils.awaitTerminationUninterruptibly;
+import static io.luna.util.concurrent.ThreadUtils.awaitTerminationUninterruptibly;
 import static org.apache.logging.log4j.util.Unbox.box;
 
 /**
