+++ util/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    `java-library`
+}
+
+group = "luna"
+version = "1.0"
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation("com.google.code.gson:gson:2.8.5")
+    implementation("com.google.guava:guava:27.0.1-jre")
+    implementation("org.apache.logging.log4j:log4j-core:2.11.1")
+    implementation("org.apache.logging.log4j:log4j-api:2.11.1")
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
