+++ util/src/main/java/io/luna/util/common/Rational.java
@@ -1,4 +1,4 @@
-package io.luna.util;
+package io.luna.util.common;
 
 import java.math.BigInteger;
 import java.util.Objects;
