+++ util/src/main/java/io/luna/util/common/ReflectionUtils.java
@@ -1,4 +1,4 @@
-package io.luna.util;
+package io.luna.util.common;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
